# analyzer/Documentation/openapi_export_20251104.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TokenLense API",
    "description": "API for accessing and visualizing model attribution data",
    "version": "1.0.0"
  },
  "paths": {
...
    "/attribution/aggregation_methods": {
      "get": {
        "tags": [
          "attribution"
        ],
        "summary": "Get Aggregation Methods",
        "description": "Get available aggregation methods",
        "operationId": "get_aggregation_methods_attribution_aggregation_methods_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AggregationOptions"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/attribution/{model}/{method}/{file_id}": {
      "get": {
        "tags": [
          "attribution"
        ],
        "summary": "Get Attribution",
        "description": "Get attribution data for a model, method, and file",
        "operationId": "get_attribution_attribution__model___method___file_id__get",
        "parameters": [
          {
            "name": "model",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model"
            }
          },
          {
            "name": "method",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Method"
            }
          },
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "File Id"
            }
          },
          {
            "name": "aggregation",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AggregationMethod",
              "description": "Aggregation method for attribution tensor",
              "default": "sum"
            },
            "description": "Aggregation method for attribution tensor"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttributionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/attribution/{model}/{method}/{file_id}/detailed": {
      "get": {
        "tags": [
          "attribution"
        ],
        "summary": "Get Detailed Attribution",
        "description": "Get detailed attribution data including token IDs and tensor information",
        "operationId": "get_detailed_attribution_attribution__model___method___file_id__detailed_get",
        "parameters": [
          {
            "name": "model",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model"
            }
          },
          {
            "name": "method",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Method"
            }
          },
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "File Id"
            }
          },
          {
            "name": "aggregation",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AggregationMethod",
              "description": "Aggregation method for attribution tensor",
              "default": "sum"
            },
            "description": "Aggregation method for attribution tensor"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttributionDetailedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/attribution/{model}/{method}/{file_id}/raw": {
      "get": {
        "tags": [
          "attribution"
        ],
        "summary": "Get Raw Attribution",
        "description": "Get raw attribution data directly from the file (for debugging)",
        "operationId": "get_raw_attribution_attribution__model___method___file_id__raw_get",
        "parameters": [
          {
            "name": "model",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model"
            }
          },
          {
            "name": "method",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Method"
            }
          },
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Raw Attribution Attribution  Model   Method   File Id  Raw Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/attribution/compare": {
      "get": {
        "tags": [
          "attribution"
        ],
        "summary": "Compare Attribution Files",
        "description": "Compare attribution data across multiple files\n\nFiles should be specified as model/method/file_id format, e.g., \"BART/attention/0\"",
        "operationId": "compare_attribution_files_attribution_compare_get",
        "parameters": [
          {
            "name": "files",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "List of file paths to compare",
              "title": "Files"
            },
            "description": "List of file paths to compare"
          },
          {
            "name": "aggregation",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AggregationMethod",
              "description": "Aggregation method",
              "default": "sum"
            },
            "description": "Aggregation method"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Compare Attribution Files Attribution Compare Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/attribution/token_importance": {
      "get": {
        "tags": [
          "attribution"
        ],
        "summary": "Get Token Importance",
        "description": "Compare importance of a specific token across multiple attribution files\n\nFiles should be specified as model/method/file_id format, e.g., \"BART/attention/0\"",
        "operationId": "get_token_importance_attribution_token_importance_get",
        "parameters": [
          {
            "name": "files",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "List of file paths to compare",
              "title": "Files"
            },
            "description": "List of file paths to compare"
          },
          {
            "name": "token_index",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Index of the token to compare",
              "title": "Token Index"
            },
            "description": "Index of the token to compare"
          },
          {
            "name": "is_target",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Whether the token is in the target (True) or source (False) sequence",
              "default": true,
              "title": "Is Target"
            },
            "description": "Whether the token is in the target (True) or source (False) sequence"
          },
          {
            "name": "aggregation",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AggregationMethod",
              "description": "Aggregation method",
              "default": "sum"
            },
            "description": "Aggregation method"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Token Importance Attribution Token Importance Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AggregationMethod": {
        "type": "string",
        "enum": [
          "sum",
          "mean",
          "l2_norm",
          "abs_sum",
          "max"
        ],
        "title": "AggregationMethod"
      },
      "AggregationOptions": {
        "properties": {
          "methods": {
            "items": {
              "$ref": "#/components/schemas/AggregationMethod"
            },
            "type": "array",
            "title": "Methods"
          },
          "default": {
            "$ref": "#/components/schemas/AggregationMethod"
          }
        },
        "type": "object",
        "required": [
          "methods",
          "default"
        ],
        "title": "AggregationOptions",
        "description": "Available aggregation methods for a given file"
      },
      "AttributionDetailedResponse": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "file_id": {
            "type": "integer",
            "title": "File Id"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "generation": {
            "type": "string",
            "title": "Generation"
          },
          "source_tokens": {
            "items": {
              "$ref": "#/components/schemas/TokenWithId"
            },
            "type": "array",
            "title": "Source Tokens"
          },
          "target_tokens": {
            "items": {
              "$ref": "#/components/schemas/TokenWithId"
            },
            "type": "array",
            "title": "Target Tokens"
          },
          "attribution_matrix": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Attribution Matrix"
          },
          "matrix_info": {
            "$ref": "#/components/schemas/AttributionMatrixInfo"
          },
          "aggregation": {
            "type": "string",
            "title": "Aggregation"
          },
          "exec_time": {
            "type": "number",
            "title": "Exec Time",
            "description": "Execution time of the attribution in seconds"
          },
          "original_attribution_shape": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Original Attribution Shape",
            "description": "Original shape of the attribution tensor before aggregation"
          }
        },
        "type": "object",
        "required": [
          "model",
          "method",
          "file_id",
          "prompt",
          "generation",
          "source_tokens",
          "target_tokens",
          "attribution_matrix",
          "matrix_info",
          "aggregation",
          "exec_time",
          "original_attribution_shape"
        ],
        "title": "AttributionDetailedResponse",
        "description": "Detailed response model for attribution data including token IDs"
      },
      "AttributionMatrixInfo": {
        "properties": {
          "shape": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Shape"
          },
          "dtype": {
            "type": "string",
            "title": "Dtype"
          },
          "is_attention": {
            "type": "boolean",
            "title": "Is Attention"
          },
          "tensor_type": {
            "type": "string",
            "title": "Tensor Type",
            "description": "One of: attention, input_x_gradient, lime, integrated_gradients, saliency, layer_gradient_x_activation"
          }
        },
        "type": "object",
        "required": [
          "shape",
          "dtype",
          "is_attention",
          "tensor_type"
        ],
        "title": "AttributionMatrixInfo",
        "description": "Information about the attribution matrix dimensions and structure"
      },
      "AttributionResponse": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "file_id": {
            "type": "integer",
            "title": "File Id"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "generation": {
            "type": "string",
            "title": "Generation"
          },
          "source_tokens": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Tokens"
          },
          "target_tokens": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Tokens"
          },
          "attribution_matrix": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Attribution Matrix"
          },
          "aggregation": {
            "type": "string",
            "title": "Aggregation"
          }
        },
        "type": "object",
        "required": [
          "model",
          "method",
          "file_id",
          "prompt",
          "generation",
          "source_tokens",
          "target_tokens",
          "attribution_matrix",
          "aggregation"
        ],
        "title": "AttributionResponse",
        "description": "Response model for attribution data"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TokenWithId": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "id",
          "token"
        ],
        "title": "TokenWithId",
        "description": "A token with its ID"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}